backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI and ML
google-generativeai==0.3.2
gTTS==2.5.0

# Audio processing
pydub==0.25.1

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.3
httpx==0.26.0
